Separating selector + function selector tests
