+++ yarn-project/foundation/src/abi/index.ts
@@ -1,6 +1,7 @@
 export * from './abi.js';
 export * from './buffer.js';
 export * from './encoder.js';
+export * from './event_selector.js';
 export * from './decoder.js';
-export * from './selector.js';
+export * from './function_selector.js';
 export * from './utils.js';
